refactor(testUpload)
Simply example usage within test.

Inclusion of the GraphQLUpload scalar within the resolver map is only necessary when merging an executable "local" subschema, as if the subschema has the original GraphqlUpload scalar defined, it will throw when attempting to serialize after it is merged into the gateway. The custom GraphqlUpload scalar can be used in that case to avoid that either within the local subschema or within the gateway.

In this case, the subschema is remote, and so the subschema custom scalar is imported with the default serialize method that does not throw.